% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffInt.R
\encoding{UTF-8}
\name{EffInt}
\alias{EffInt}
\title{Effective Interest rate (Engineering Economics)}
\usage{
EffInt(
  r,
  frequency = c("annual", "semiannual", "quarter", "bimonth", "month", "daily")
)
}
\arguments{
\item{r}{numeric vector that contains the nominal interest rate(s) per
period as a percent}

\item{frequency}{character vector that contains the frequency used to
obtain the number of periods [annual (1), semiannual (2), quarter (4),
bimonth (6), month (12), daily (365)]}
}
\value{
EffInt numeric vector that contains the effective interest rate
   rounded to 2 decimal places (this is the \code{i} used in the other
   Engineering Economics functions)
}
\description{
Computes the effective interest rate given the nominal interest rate per
period
}
\details{
i is expressed as

	\deqn{i = \left(1 + \frac{r}{n}\right)^n - 1}

\describe{
\item{\emph{i}}{the "effective interest rate per interest period"}
\item{\emph{r}}{the "nominal interest rate"}
\item{\emph{n}}{the "number of compounding periods per year"}
}
}
\examples{

library(iemisc)

# Example 4-28 from Sullivan Reference text (page 165)
EffInt(1.375, frequency = "month")
# the nominal interest rate per period (month) is 1.375\\%


# Example from SFPE Reference text
EffInt(18 / 12, frequency = "month")
# the nominal interest rate is 18\\% per year or 18\\% / 12 months


}
\references{
\enumerate{
 \item \emph{SFPE Handbook of Fire Protection Engineering}. 3rd Edition, DiNenno, P. J.; Drysdale, D.; Beyler, C. L.; Walton, W. D., Editor(s), page 5-95, 2002. Chapter 7; Section 5; NFPA HFPE-02. See \url{https://web.archive.org/web/20180127185316/http://fire.nist.gov/bfrlpubs/build02/PDF/b02155.pdf}.
 \item William G. Sullivan, Elin M. Wicks, and C. Patrick Koelling, \emph{Engineering Economy}, Fourteenth Edition, Upper Saddle River, New Jersey: Pearson/Prentice Hall, 2009, pages 164-165.
}
}
\author{
Irucka Embry
}
