% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\encoding{UTF-8}
\name{f2}
\alias{f2}
\title{Moody Equation}
\usage{
f2(eps, Re, D = NULL)
}
\arguments{
\item{eps}{numeric vector that contains the specific roughness for the pipe
material (m or ft)}

\item{Re}{numeric vector that contains the Reynolds number (dimensionless)}

\item{D}{numeric vector that contains the inner diameters of the pipe (m or
ft)}
}
\value{
the dimensionless Darcy friction factor (f) as a numeric \code{\link[base]{vector}}
}
\description{
Moody Equation
}
\note{
Note: Please refer to iemisc: Calculating the Friction Loss Examples vignette
for the examples
}
\seealso{
\code{\link{f1}} for the additional seealso, description, details, and references sections, \code{\link{f3}}, \code{\link{f4}}, \code{\link{f5}}, \code{\link{f6}}, \code{\link{f7}}, \code{\link{f8}}
}
\author{
Irucka Embry
}
