% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractdiff.R
\encoding{UTF-8}
\name{fractdiff}
\alias{fractdiff}
\title{Fractional Differences (GNU Octave/MATLAB compatible)}
\source{
\enumerate{
   \item r - How to not run an example using roxygen2? - Stack Overflow answered and edited by samkart on Jul 9 2017. (Also see the additional comments in response to the answer.) See \url{https://stackoverflow.com/questions/12038160/how-to-not-run-an-example-using-roxygen2}.
   \item devtools - Issues in R package after CRAN asked to replace dontrun by donttest - Stack Overflow answered by Hong Ooi on Sep 1 2020. (Also see the additional comments in response to the answer.) See \url{https://stackoverflow.com/questions/63693563/issues-in-r-package-after-cran-asked-to-replace-dontrun-by-donttest}.
}
}
\usage{
fractdiff(x, d)
}
\arguments{
\item{x}{A numeric vector}

\item{d}{A numeric scalar}
}
\value{
Return the fractional differences.
}
\description{
"Compute the fractional differences (1 - L)^d * x where L denotes the lag-
operator and d is greater than -1." Source: Eaton page 853.
}
\note{
Note: Please refer to the iemisc: Examples from GNU Octave Rem, Mod, and
fractdiff Compatible Functions vignette for an example
}
\references{
John W. Eaton, David Bateman, Søren Hauberg, and Rik Wehbring (November 2022). \emph{GNU Octave: A high-level interactive language for numerical computations: Edition 7 for Octave version 7.3.0}. \url{https://docs.octave.org/octave.pdf}. Page 853.
}
\author{
Irucka Embry

Irucka Embry, GNU Octave Developers (GNU Octave code)
}
