% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitcomma.R
\encoding{UTF-8}
\name{\%qsin\%}
\alias{\%qsin\%}
\title{Quick Search}
\source{
data.table - R case-insensitive \%in\% - Stack Overflow answered and edited by Dylan Russell on Sep 13, 2020. See \url{https://stackoverflow.com/questions/63874824/r-case-insensitive-in}.
}
\usage{
string \%qsin\% vector
}
\arguments{
\item{string}{character vector that contains the values to match}

\item{vector}{character vector that has the values to be matched against}
}
\value{
logical vector the length of the original string
}
\description{
This function performs a quick, case insensitive search of strings
}
\details{
'Utilizes chin from data.table to quickly complete a case
   insensitive search through a character vector to return a logical vector
   of string detections. Will always return TRUE or FALSE for each position
   of string regardless of NA missing values in either provided vector. NA
   in string will never match an NA value in the vector.'
}
\examples{

# Examples

x <- c("apple", "banana", "cherry", NA)

"apple" \%qsin\% x

c("APPLE", "BANANA", "coconut", NA) \%qsin\% x




}
\author{
Dylan Russell (stackoverflow code, examples, and function details, function description), Irucka Embry
}
