% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemisc-package.R, R/iemisc.R
\docType{package}
\name{iemisc-package}
\alias{iemisc-package}
\title{iemisc: Irucka Embry's miscellaneous functions}
\description{
iemisc provides many useful functions. There are statistical analysis [RMS,
approximate and relative error, maximum mean relative error (MAXRE), and
mean relative error (MRE), coefficient of variation (CV), range, harmonic
mean, geometric mean], engineering economics (benefit-cost, future value,
present value, annual value, gradients, interest, periods, etc.), geometry
(polygon area, sphere volume, and right triangle validation), civil &
environmental/water resources engineering [Concrete Mix Design for Normal
Strength & Structural Lightweight Concrete, Construction Measurements with
and without Fractions, Engineering Surveying Calculations, Air Stripping,
Saturated Vapor Pressure for Ice and Water, Density of Water, Dynamic and
Kinematic Viscosity of Water, Specific Gravity and Volume, Unit Weight,
Surface Tension of Water, Weighted C Factor, Weighted Curve Number, Darcy
friction factor (f), the Reynolds number, Manning's n,
Gauckler-Manning-Strickler equations for open channel flow], quick search,
return character vectors in order, mortality rate calculations, proportion
solver, sum of all digits in a vector to a single integer, string
manipulation, a version of linear interpolation for use with NAs, Python
compatible floor division function, GNU Octave/MATLAB compatible
trigonometric functions in degrees, GNU Octave/MATLAB compatible section
properties, remainder, modulus, number of dimensions, row vector and column
vector tests, fractional differences, size, numel, and length functions.

iemisc provides many useful functions. There are statistical analysis [RMS,
approximate and relative error, maximum mean relative error (MAXRE), and
mean relative error (MRE), coefficient of variation (CV), range, harmonic
mean, geometric mean], engineering economics (benefit-cost, future value,
present value, annual value, gradients, interest, periods, etc.), geometry
(polygon area, sphere volume, and right triangle validation), civil &
environmental/water resources engineering [Concrete Mix Design for Normal
Strength & Structural Lightweight Concrete, Construction Measurements with
and without Fractions, Engineering Surveying Calculations, Air Stripping,
Saturated Vapor Pressure for Ice and Water, Density of Water, Dynamic and
Kinematic Viscosity of Water, Specific Gravity and Volume, Unit Weight,
Surface Tension of Water, Weighted C Factor, Weighted Curve Number, Darcy
friction factor (f), the Reynolds number, Manning's n,
Gauckler-Manning-Strickler equations for open channel flow], quick search,
return character vectors in order, mortality rate calculations, proportion
solver, sum of all digits in a vector to a single integer, string
manipulation, a version of linear interpolation for use with NAs, Python
compatible floor division function, GNU Octave/MATLAB compatible
trigonometric functions in degrees, GNU Octave/MATLAB compatible section
properties, remainder, modulus, number of dimensions, row vector and column
vector tests, fractional differences, size, numel, and length functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/iembry/iemisc}
  \item Report bugs at \url{https://gitlab.com/iembry/iemisc/-/issues}
}


Useful links:
\itemize{
  \item \url{https://gitlab.com/iembry/iemisc}
  \item Report bugs at \url{https://gitlab.com/iembry/iemisc/-/issues}
}

}
\author{
\strong{Maintainer}: Irucka Embry \email{iembry@ecoccs.com}

Authors:
\itemize{
  \item Felix Andrews (zoo code) [contributor]
  \item Hans Werner Borchers (pracma code) [contributor]
  \item Samit Basu (FreeMat code) [contributor]
  \item David Bateman (GNU Octave code) [contributor]
  \item Rik Wehbring (GNU Octave code) [contributor]
  \item Didier Clamond (MATLAB code) [contributor]
  \item Berry Boessenkool (checkLL and degree functions from OSMscale) [contributor]
  \item Tyler Rinker (lookup and lookup_helper functions from qdapTools) [contributor]
  \item Colin Caprani (secprop MATLAB function) [contributor]
  \item Teodor Ciuraru (stackoverflow R code) [contributor]
  \item Dylan Russell (stackoverflow R code) [contributor]
  \item John Wallace (stackoverflow R code) [contributor]
  \item kaijagahm (R bloggers R code) [contributor]
  \item Colin B. Macdonald (OctSymPy code) [contributor]
  \item John D Page (JavaScript code) [contributor]
  \item Josh O'Brien (stackoverflow R code) [contributor]
  \item R. van Twisk (LibreCAD code) [contributor]
}

}
\keyword{internal}
