% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupQT.R
\name{lookupQT}
\alias{lookupQT}
\title{Hash Table/Dictionary Lookup
 
These functions were originally contained in "qdapTools" version 1.3.4
\code{lookup} - \href{http://datatable.r-forge.r-project.org/}{\pkg{data.table}} 
based hash table useful for large vector lookups.}
\usage{
lookupQT(terms, key.match, key.reassign = NULL, missing = NA)
}
\arguments{
\item{terms}{A vector of terms to undergo a lookup.}

\item{key.match}{Takes one of the following: (1) a two column data.frame of a 
match key and reassignment column, (2) a named list of vectors (Note: if 
data.frame or named list supplied no key reassign needed) or (3) a single 
vector match key.}

\item{key.reassign}{A single reassignment vector supplied if key.match is 
not a two column data.frame/named list.}

\item{missing}{Value to assign to terms not matching the key.match.  If set 
to \code{NULL} the original values in \code{terms} corresponding to the 
missing elements are retained.}
}
\value{
Outputs A new vector with reassigned values.
}
\description{
Hash Table/Dictionary Lookup
 
These functions were originally contained in "qdapTools" version 1.3.4
\code{lookup} - \href{http://datatable.r-forge.r-project.org/}{\pkg{data.table}} 
based hash table useful for large vector lookups.
}
\author{
Tyler Rinker ('qdapTools' package version 1.3.4)
}
