% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rem.R
\encoding{UTF-8}
\name{Mod_octave}
\alias{Mod_octave}
\title{Modulus Operation (GNU Octave/MATLAB compatible)}
\usage{
Mod_octave(x, n)
}
\arguments{
\item{x, n}{An R object (array, matrix, vector)}
}
\value{
Return the computed modulo of x and y.
}
\description{
Obtain the modulo of x and y in a manner compatible with GNU Octave/MATLAB.
}
\examples{

# Example from FreeMat

library(iemisc)

Mod_octave(18, 12)

# Please refer to the iemisc: Examples from GNU Octave Rem, Mod, and
# fractdiff Compatible Functions vignette for additional examples





}
\references{
\enumerate{
   \item Samit Basu (2002-2006). FreeMat v4.0, \url{https://freemat.sourceforge.net/help/mathfunctions_mod.html}.
   \item John W. Eaton, David Bateman, Søren Hauberg, and Rik Wehbring (November 2022). \emph{GNU Octave: A high-level interactive language for numerical computations: Edition 7 for Octave version 7.3.0}. \url{https://docs.octave.org/octave.pdf}. Page 564.
}
}
\author{
Irucka Embry, Samit Basu (FreeMat)
}
