% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secprop.R
\encoding{UTF-8}
\name{secprop}
\alias{secprop}
\title{Section Properties Calculator [GNU Octave/MATLAB]}
\source{
loops - r - foreach unable to find object within function - Stack Overflow answered by sumshyftw on Jun 7 2019. See \url{https://stackoverflow.com/questions/56498824/r-foreach-unable-to-find-object-within-function/56499056}.
}
\usage{
secprop(
  outer_coord,
  inner_coord,
  original_plot = c(0, 1),
  final_plot = c(0, 1)
)
}
\arguments{
\item{outer_coord}{numeric matrix that contains the "outer coordinates (x, y)"
Reference: Caprani}

\item{inner_coord}{numeric matrix that contains the "coordinates for a void"
Reference: Caprani}

\item{original_plot}{integer vector that contains 0, 1 only. 0 represents
do not print the original plot and 1 is for printing the original plot. The
default value is 0 (no original plot).}

\item{final_plot}{integer vector that contains 0, 1 only. 0 represents do
not print the final (transformed) plot and 1 is for printing the final
(transformed) plot. The default value is 1 (final plot will be shown).}
}
\value{
SP numeric vector that contains "A (Area), I (Second Moment of Area),
  yt and yb properties of the section."
  Reference: Caprani
}
\description{
This function "is a short but really useful script that calculates the
section properties for an arbitrary shape with holes."
Reference: Caprani
}
\note{
Note: Please refer to the iiemisc: secprop Example (R and GNU Octave)
vignette for the examples
}
\references{
Colin Caprani, "Section Properties Calculator", \url{https://www.colincaprani.com/programming/matlab/}.
}
\author{
Colin Caprani (secprop MATLAB function), Irucka Embry (secprop R function)
}
