% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemisc.R
\docType{package}
\name{iemisc}
\alias{iemisc}
\alias{iemisc-package}
\title{iemisc: Irucka Embry's miscellaneous functions}
\description{
iemisc provides many useful functions. There are statistical analysis [RMS,
coefficient of variation (CV), approximate and relative error, range,
harmonic mean, geometric mean], engineering economics (benefit-cost,
future value, present value, annual value, gradients, interest, periods,
etc.), geometry (sphere volume and right triangle), civil & environmental/
water resources engineering (Normal Strength Concrete Mix Design, Manning's
n, Gauckler-Manning-Strickler equation for geometric cross-sections), a
version of linear interpolation for use with NAs, & GNU Octave/MATLAB
compatible size, numel, and length_octave functions.
}

